feat(nixrs): Make store traits have a StoreDir
This implements a HasStoreDir trait that any DaemonStore and LocalDaemonStore
must implement. This is so that when you have a store you also have access to
the StoreDir relevant for that store.

It also removes that store_dir functions from NixRead and NixWrite and make
those require this trait instead.