feat(capnp-rpc-tokio): Allow better control client shutdown
This makes the Cap'n Proto client builder return a new ShutdownClient
that allows more control over how the rpc client connnection is shut
down. The old behavior where when the client is dropped the connection
is also shut down is still supported but now requires a call to an
`.into_client()` function to return such a client.