chore(nixrs): Add note about flaky test
One of the tests for the nar reader sometimes stalls and never finishes.
I have seen this several times but it is a non-reproducable timing issue
so I have made a note of the exact test that failed to verify that
it is that exact test and arguments that fails and not several tests.