feat(nixrs-capnp): Add CapLookup helpers
This adds helpers to make it easier to deal with the CapLookup
interface.

It contains a wrapper for the client that makes it easier to lookup
a Cap'n Proto interface.

And it has a registry helper that implement CapLookup and that you can
use to add clients that can be looked up.