refactor(nixrs): Rearrange daemon types to better modules
This moves and rearranges some of the daemon types to more appropriate
modules.

Specifically it gets rid of the `daemo::wire::types2` module and
incorporates the types from there into `daemon::types` and
`daemon::wire::types` so that `daemon::types` is for using the daemon
protocol as a client or server and `daemon::wire` is for implementing the
daemon protocol and we might even only make this module public under the
`internal` feature.

This commit also moves test related daemon code to `test::daemon` and
`test::arbitrary::daemon` and gets rid of the public `test::daemon::mock`
module.