refactor(nixrs): Get rid of nixrs-derive feature
The `nixrs-derive` feature was always kind of weird the no clear purpose
for what a user of nixrs would want that feature for. So I have removed it
and instead now use the `daemon` feature exclusively for when the daemon
protocol code is enabled. The `daemon-serde` feature still exists but that
is more an internal feature to get around a circular dependency isssue when
compiling with Nix.